- patches.fixes/patch-2.6.11-rc1: 2.6.11-rc1. - patches.fixes/patch-2.6.11-rc1-bk8: 2.6.11-rc1-bk8.
suse-commit: d7f9ab326b9784ffceda85742a58dc43d7f28248

+++ arch/um/drivers/cow.h
@@ -21,11 +21,12 @@ extern int file_reader(__u64 offset, char *buf, int len, void *arg);
 extern int read_cow_header(int (*reader)(__u64, char *, int, void *),
 			   void *arg, __u32 *version_out,
 			   char **backing_file_out, time_t *mtime_out,
-			   __u64 *size_out, int *sectorsize_out,
+			   unsigned long long *size_out, int *sectorsize_out,
 			   __u32 *align_out, int *bitmap_offset_out);
 
 extern int write_cow_header(char *cow_file, int fd, char *backing_file,
-			    int sectorsize, int alignment, long long *size);
+			    int sectorsize, int alignment,
+			    unsigned long long *size);
 
 extern void cow_sizes(int version, __u64 size, int sectorsize, int align,
 		      int bitmap_offset, unsigned long *bitmap_len_out,
